+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+### Fixed
+
+- Schema type checks fail on nested fields for Databricks spark (#618)
+
 ## [0.10.21] - 2025-02-06
 
 ### Added