+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Schema type checks fail on nested fields for Databricks spark (#618)
 
+### Removed
+
+- `datacontract test --examples`: This option was removed as it was not very popular and top-level
+  examples section is deprecated in the Data Contract Specification v1.1.0 (#628)
+
 ## [0.10.21] - 2025-02-06
 
 ### Added